# services/orca-core/src/main/resources/openapi/orca-core.yaml
openapi: 3.0.3
...
info:
  title: orca-core
  description: |
    The world as configured (§C1) — tenancy, sites and lanes, users, devices, design artifacts, custom entities, licence verification.
  version: 0.0.1
servers:
  - url: http://localhost:8081
    description: Local development
...
paths:
...
  /api/v1/roles:
    get:
      tags: [roles]
      operationId: listRoles
      summary: List roles, each with its site scope and resolved entitlement grants
      responses:
        "200":
          description: The roles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RolesEnvelope"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
    post:
      tags: [roles]
      operationId: createRole
      summary: Create a role, optionally with grants and site scope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleEnvelope"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "409":
          description: An active role already carries that name
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "422":
          description: An entitlement code or site is unknown
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
  /api/v1/roles/{roleExternalId}:
    patch:
      tags: [roles]
      operationId: updateRole
      summary: Update a role, its grants or its site scope
      description: |
        Absent fields are unchanged. `entitlementCodes` and `siteExternalIds`,
        when present, REPLACE the grant set and the site scope respectively —
        they are declarative, not additive. A role still held by an active user
        cannot be retired.
      parameters:
        - name: roleExternalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleRequest"
      responses:
        "200":
          description: The role after the change
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleEnvelope"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "404":
          description: No such role
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "409":
          description: An active role already carries that name, or active users still hold this role
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "422":
          description: An entitlement code or site is unknown
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
...
  /api/v1/resource-configurations/{scopeType}/{resourceExternalId}:
    get:
      tags: [resourceConfigurations]
      operationId: readResourceConfiguration
      summary: The custom variables of one scoped resource
      parameters:
        - name: scopeType
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ResourceScopeType"
        - name: resourceExternalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The variables
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceConfigurationEnvelope"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "404":
          description: No such resource at this installation
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
    put:
      tags: [resourceConfigurations]
      operationId: replaceResourceConfiguration
      summary: Replace the custom variables of one scoped resource
      description: Declarative — the request's set is the whole set for that resource.
      parameters:
        - name: scopeType
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ResourceScopeType"
        - name: resourceExternalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/ResourceConfigurationEntry"
      responses:
        "200":
          description: The variables after the change
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceConfigurationEnvelope"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "400":
          description: The request repeats an entry
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "404":
          description: No such resource at this installation
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
  /api/v1/settings:
    get:
      tags: [settings]
      operationId: listSettings
      summary: Every known key, its metadata and its current value
      responses:
        "200":
          description: The registry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsEnvelope"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
  /api/v1/settings/{settingKey}:
    put:
      tags: [settings]
      operationId: writeSetting
      summary: Write one setting
      description: |
        Validated against the registry — an unknown key is refused, a value
        that fails the key's type is refused, and a SECRET-SHAPED key is
        refused outright with its own code before the registry is even
        consulted (1.x stored SMTP passwords and client secrets in this table;
        2.0 never will). Every accepted write appends to the history.
      parameters:
        - name: settingKey
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WriteSettingRequest"
      responses:
        "200":
          description: The setting after the write
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingEnvelope"
        "400":
          description: The value fails the key's type
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "422":
          description: Unknown key (SETTING_UNKNOWN), or a secret-shaped key (SETTING_SECRET_REJECTED)
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
...
components:
  schemas:
...
    RoleSummary:
      type: object
      required: [externalId, name, retired, siteExternalIds, entitlementCodes]
      properties:
        externalId:
          type: string
        name:
          type: string
        description:
          type: string
        siteExternalIds:
          type: array
          description: The sites this role is scoped to (role-based site scoping — 1.x user↔site mapping is dead).
          items:
            type: string
        entitlementCodes:
          type: array
          description: The granted catalog leaves, by stable code.
          items:
            type: string
        retired:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateRoleRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 3000
        siteExternalIds:
          type: array
          items:
            type: string
        entitlementCodes:
          type: array
          items:
            type: string
    UpdateRoleRequest:
      type: object
      description: |
        Absent fields are unchanged; `entitlementCodes` and `siteExternalIds`
        REPLACE their sets when present.
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 3000
        siteExternalIds:
          type: array
          items:
            type: string
        entitlementCodes:
          type: array
          items:
            type: string
        retired:
          type: boolean
    RoleEnvelope:
      allOf:
        - $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/RoleSummary"
    RolesEnvelope:
      allOf:
        - $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/RoleSummary"

    # --- WP2 · teams & templates --------------------------------------------
...
    ResourceScopeType:
      type: string
      enum: [SITE, AREA, LANE, DEVICE]
...
    ResourceConfigurationEntry:
      type: object
      required: [key]
      properties:
        key:
          type: string
          maxLength: 200
        value:
          type: string
          maxLength: 2000
    ResourceConfigurationView:
      type: object
      required: [scopeType, resourceExternalId, entries]
      properties:
        scopeType:
          $ref: "#/components/schemas/ResourceScopeType"
        resourceExternalId:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/ResourceConfigurationEntry"
    ResourceConfigurationEnvelope:
      allOf:
        - $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ResourceConfigurationView"

    # --- WP4 · settings · workspace · audit ---------------------------------
    SettingSummary:
      type: object
      required: [key, valueType]
      properties:
        key:
          type: string
        valueType:
          type: string
          enum: [STRING, INTEGER, BOOLEAN]
        description:
          type: string
        defaultValue:
          type: string
        value:
          type: string
          description: The current value, absent when never written (the default applies).
        updatedBy:
          type: string
        updatedAt:
          type: string
          format: date-time
    WriteSettingRequest:
      type: object
      required: [value]
      properties:
        value:
          type: string
          maxLength: 2000
    SettingEnvelope:
      allOf:
        - $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/SettingSummary"
    SettingsEnvelope:
      allOf:
        - $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/SettingSummary"
...